aion_taxonomy/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aion-taxonomy"
version = "1.0.0"
description = "Rule-based event classification and impact scoring for Indian equity markets"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.9"
authors = [
    {name = "AION Contributors", email = "aion@opensource.example.com"}
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial :: Investment",
    "Typing :: Typed",
]
dependencies = [
    "pyyaml>=6.0",
    "pandas>=2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "mypy>=1.0",
]
sectormap = [
    "aion-sectormap>=1.0",
]
sentiment = [
    "aion-sentiment-in>=2.0",
]

[project.urls]
Homepage = "https://github.com/aion-open-source/aion_taxonomy"
Documentation = "https://github.com/aion-open-source/aion_taxonomy#readme"
Repository = "https://github.com/aion-open-source/aion_taxonomy.git"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
aion_taxonomy = ["data/*.yaml", "data/*.json"]

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312', 'py313']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
addopts = "-v --cov=aion_taxonomy --cov-report=term-missing"
